Crate minerva_voucher

A compact CBOR-encoded voucher defined by Constrained BRSKI.

Examples (!! WIP !!)

use minerva_voucher::{Voucher, Sign, Validate};

// (Add notes on the PSA crypto context ...)
#[cfg(feature = "v3")]
init_psa_crypto();

let mut vrq = Voucher::new_vrq();

// ...

A Voucher with a known list of attributes can be initialized using macros:

use minerva_voucher::{vch, vrq};

let vch = vch![
    Attr::Assertion(Assertion::Logged),
    Attr::SerialNumber("00-11-22-33-44-55".as_bytes().to_vec())];

let vrq = vrq![
    Attr::Assertion(Assertion::Proximity),
    Attr::SerialNumber("00-11-22-33-44-55".as_bytes().to_vec())];

A raw CBOR-encoded voucher can be decoded into a Voucher through the TryFrom and/or TryInto traits.

use core::convert::{TryFrom, TryInto};

assert!(Voucher::try_from(VCH_JADA).is_ok());

let vch: Voucher = VCH_JADA.try_into().unwrap();

vch.iter().for_each(|attr| {
    println!("attr: {:?}", attr);
    match attr {
        Attr::Assertion(x) => assert_eq!(x, &Assertion::Proximity),
        Attr::CreatedOn(x) => assert_eq!(x, &1475868702),
        Attr::ExpiresOn(x) => assert_eq!(x, &1506816000),
        Attr::Nonce(x) => assert_eq!(x, &[97, 98, 99, 100, 49, 50, 51, 52, 53]),
        Attr::PinnedDomainPubk(x) => assert_eq!(x[0..4], [77, 70, 107, 119]),
        Attr::SerialNumber(x) => assert_eq!(x, "JADA123456789".as_bytes()),
        _ => panic!(),
    }
});

Modules

attr

debug

Macros

vch

todo

vrq

todo

Structs

Voucher

Enums

SignatureAlgorithm

An enum identifying supported signature algorithms. Currently ES256 (ECDSA with P256 and SHA256), ES384 (ECDSA with P384 and SHA384) ES512 (ECDSA with P521 and SHA512), and PS256 (RSASSA-PSS with SHA256) are supported. Note that with PS256, the salt length is defined to be 32 bytes.

VoucherError

Traits

Sign

Validate