Module minerva_voucher::debug

Enums

CborType

Functions

content_comp

content_comp_permissive

decode

Read the CBOR structure in bytes and return it as a CborType. To prevent stack exhaustion, the maximum nested depth of CBOR objects (for example, an array of an array of an array…) is 256. If the input data describes a CBOR structure that exceeds this limit, an error will be returned.

sig_one_struct_bytes_from