Struct minerva_voucher::Voucher

pub struct Voucher { /* private fields */ }

Implementations

impl Voucher

pub fn new_vrq() -> Self

pub fn new_vch() -> Self

pub fn is_vrq(&self) -> bool

pub fn is_vch(&self) -> bool

pub fn get(&self, adisc: AttrDisc) -> Option<&Attr>

Returns a reference to the attribute in the voucher, if any, that corresponds to the given attribute discriminant value.

Examples

use minerva_voucher::{Voucher, attr::*};

let mut vrq = Voucher::new_vrq();
vrq.set(Attr::CreatedOn(1475868702));

assert_eq!(vrq.get(ATTR_CREATED_ON), Some(&Attr::CreatedOn(1475868702)));
assert_eq!(vrq.get(ATTR_SERIAL_NUMBER), None);

pub fn set(&mut self, attr: Attr) -> &mut Self

Adds an attribute to the voucher, replacing the existing attribute, if any, that corresponds to the given one. Returns a mut reference to the voucher.

Examples

use minerva_voucher::{Voucher, attr::*};

let mut vrq = Voucher::new_vrq();
assert_eq!(vrq.get(ATTR_CREATED_ON), None);

vrq.set(Attr::CreatedOn(1475868702));
assert_eq!(vrq.get(ATTR_CREATED_ON), Some(&Attr::CreatedOn(1475868702)));

vrq.set(Attr::CreatedOn(1599086034));
assert_eq!(vrq.get(ATTR_CREATED_ON), Some(&Attr::CreatedOn(1599086034)));

pub fn remove(&mut self, adisc: AttrDisc) -> bool

Removes an attribute from the voucher. Returns whether the attribute was present in the voucher.

Examples

use minerva_voucher::{Voucher, attr::*};

let mut vrq = Voucher::new_vrq();
vrq.set(Attr::CreatedOn(1475868702));

assert_eq!(vrq.remove(ATTR_CREATED_ON), true);
assert_eq!(vrq.remove(ATTR_CREATED_ON), false);

pub fn take(&mut self, adisc: AttrDisc) -> Option<Attr>

Removes and returns the attribute in the voucher, if any, that corresponds to the given attribute discriminant value.

Examples

use minerva_voucher::{Voucher, attr::*};

let mut vrq = Voucher::new_vrq();

vrq.set(Attr::CreatedOn(1475868702));
assert_eq!(vrq.take(ATTR_CREATED_ON), Some(Attr::CreatedOn(1475868702)));
assert_eq!(vrq.take(ATTR_CREATED_ON), None);

let sn = b"00-D0-E5-F2-00-02";
vrq.set(Attr::SerialNumber(sn.to_vec()));
assert_eq!(vrq.take(ATTR_SERIAL_NUMBER), Some(Attr::SerialNumber(sn.to_vec())));
assert_eq!(vrq.take(ATTR_SERIAL_NUMBER), None);

pub fn len(&self) -> usize

Returns the number of attributes in the voucher.

Examples

use minerva_voucher::{Voucher, attr::Attr};

let mut vrq = Voucher::new_vrq();
assert_eq!(vrq.len(), 0);
vrq.set(Attr::CreatedOn(1475868702));
assert_eq!(vrq.len(), 1);

pub fn iter(&self) -> impl Iterator<Item = &Attr> + '_

Gets an iterator that visits the attributes in the voucher.

Examples

use minerva_voucher::{Voucher, attr::{Attr, Assertion}};

let mut vrq = Voucher::new_vrq();

vrq.set(Attr::Assertion(Assertion::Proximity))
    .set(Attr::CreatedOn(1599086034))
    .set(Attr::SerialNumber(b"00-D0-E5-F2-00-02".to_vec()));

let mut vrq_iter = vrq.iter();
assert!(vrq_iter.next().is_some());
assert!(vrq_iter.next().is_some());
assert!(vrq_iter.next().is_some());
assert!(vrq_iter.next().is_none());

pub fn get_signature(&self) -> (&[u8], &SignatureAlgorithm)

…

Examples

;

pub fn serialize(&self) -> Result<Vec<u8>, VoucherError>

…

Examples

;

pub fn get_signer_cert(&self) -> Option<&[u8]>

Returns a reference to the signer certificate in the voucher, if any.

Examples

use minerva_voucher::Voucher;

let mut vrq = Voucher::new_vrq();

assert_eq!(vrq.get_signer_cert(), None);
vrq.set_signer_cert(&[4, 186, 197, 177, 28, 173, 143, 153, 249, 199, 43, 5, 207, 75, 158, 38, 210, 68, 220, 24, 159, 116, 82, 40, 37, 90, 33, 154, 134, 214, 160, 158, 255, 32, 19, 139, 248, 45, 193, 182, 213, 98, 190, 15, 165, 74, 183, 128, 74, 58, 100, 182, 215, 44, 207, 237, 107, 111, 182, 237, 40, 187, 252, 17, 126]);
assert_eq!(vrq.get_signer_cert().unwrap().len(), 65);

pub fn set_signer_cert(&mut self, cert: &[u8]) -> &mut Self

Adds a singer certificate to the voucher.

Examples

use minerva_voucher::Voucher;

let mut vrq = Voucher::new_vrq();

assert_eq!(vrq.get_signer_cert(), None);
vrq.set_signer_cert(&[4, 186, 197, 177, 28, 173, 143, 153, 249, 199, 43, 5, 207, 75, 158, 38, 210, 68, 220, 24, 159, 116, 82, 40, 37, 90, 33, 154, 134, 214, 160, 158, 255, 32, 19, 139, 248, 45, 193, 182, 213, 98, 190, 15, 165, 74, 183, 128, 74, 58, 100, 182, 215, 44, 207, 237, 107, 111, 182, 237, 40, 187, 252, 17, 126]);
assert_eq!(vrq.get_signer_cert().unwrap().len(), 65);

pub fn to_sign(&mut self, alg: SignatureAlgorithm) -> (&mut Vec<u8>, &[u8])

Interfaces with meta data to be used in ECDSA based signing.

Examples

;

pub fn to_validate(&self) -> (Option<&[u8]>, &[u8], &SignatureAlgorithm, &[u8])

Interfaces with meta data to be used in ECDSA based validation.

Examples

;

pub fn dump(&self)

…

Examples

;